requirements.txt
Django==4.2
psycopg2-binary==2.9.5
django-filter==23.3
Pillow==9.4.0
django-resized==1.0.2